pyproject.toml
[project]
name = "microtasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "faker>=37.4.2",
    "ruff>=0.12.4",
]
